Fix some Cvars Using Double as Their Type (#2333)
The engine [does not technically
support](https://github.com/user-attachments/assets/9cb5645b-b767-4f1e-aa43-6fae3ee4e9c8)
the `double` type, but it seems some system was converting the
interpreted string into a double so CVars worked fine when defined by a
config file or when using defaults, but the `cvar` *command* doesn't
work with it. I'd add support for it, but we haven't switched engines
yet.

---

<details><summary><h1>Media</h1></summary>
<p>

Before

![image](https://github.com/user-attachments/assets/c2f760a9-705f-4e68-a9da-5c837abac40c)

After

![image](https://github.com/user-attachments/assets/e7f73ac3-9adb-4440-8667-52b41c5bdbc2)

</p>
</details>

+++ Content.Shared/CCVar/CCVars.cs
@@ -273,16 +273,16 @@ public sealed partial class CCVars : CVars
     public static readonly CVarDef<bool> MOTDBuletinEnable =
         CVarDef.Create("chat.motd_buletin_enable", true, CVar.SERVER | CVar.REPLICATED | CVar.ARCHIVE, "If the MOTD buletin appears in lobby");
 
-        #region GhostRespawn
+    #region GhostRespawn
 
-        public static readonly CVarDef<double> GhostRespawnTime =
-            CVarDef.Create("ghost.respawn_time", 15d, CVar.SERVERONLY);
+    public static readonly CVarDef<float> GhostRespawnTime =
+        CVarDef.Create("ghost.respawn_time", 15f, CVar.SERVERONLY);
 
-        public static readonly CVarDef<int> GhostRespawnMaxPlayers =
-            CVarDef.Create("ghost.respawn_max_players", 40, CVar.SERVERONLY);
+    public static readonly CVarDef<int> GhostRespawnMaxPlayers =
+        CVarDef.Create("ghost.respawn_max_players", 40, CVar.SERVERONLY);
 
-        public static readonly CVarDef<bool> GhostAllowSameCharacter =
-            CVarDef.Create("ghost.allow_same_character", false, CVar.SERVERONLY);
+    public static readonly CVarDef<bool> GhostAllowSameCharacter =
+        CVarDef.Create("ghost.allow_same_character", false, CVar.SERVERONLY);
 
-        #endregion
+    #endregion
 }
